Backend/requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
python-docx
openai
langchain
langchain-openai
langchain-community
azure-search-documents
azure-identity